Integrate CR / CRLF replacement for streaming.
HTML spec defines it at preprocessing stage, which is what I've done before, but this doesn't play well with streaming unless you want to introduce a bunch of hacks plus keep a separate pass just for such replacement before feeding the actual tokenizer.

html5lib tests don't cover all of the places where CR LF? might appear, but hopefully I haven't missed any; if I did, that should be easy to fix in the future and in any case this is non-critical replacement as absence of it will result in inconsistent line endings (the only place where it's critical was TagNameSpace which is already covered).

In SafeStrings and Data it theoretically could be done more efficiently by looking ahead to the character after CR and, if it's LF, including it as start of the next slice and not character on its own, but in practice I've tried to do this for the last two days and every time it resulted in a messed up state machines with some missed edge cases, bigger output file and, consequently, slower tokenization, so implemented it in a simpler way for those states; instead, Comment and CData states are already implemented as state charts due to their complexity, so the initial approach is used there.