Get rid of non-streaming-friendly AppendSliceAfterTheMark.
Adds char-by-char streaming comparison of "special" (appropriate) end tag names instead of old collect-and-compare approach.

This requires more actions, but allows to get rid of AppendSliceAfterTheMark so that only AppendSlice and AppendSliceBeforeTheMark are left, which can be safely executed on the end of chunk to enable continuous streaming of text instead of nested buffering.