Remove confusing LHTML_TAG_UNKNOWN
It claims to be for custom elements, but in fact 0 is generated only when the hash doesn't fit into uint64.

Custom elements that do fit will still produce a non-zero hash that is just not listed in the enum, so it's best to remove this one so that no one would rely on it.