Simplify Character definition
"Unsafe null" and "entities allowed" are technically properties of the current context and not specific to the token produced.

For example, if some transformer produces a character token when processing contents between <script>...</script>, such character token will be subject to the same processing rules by the browser as if it would be parsed from the original text.

Also, this makes producing tokens more convenient as you don't need to worry about the .kind anymore.

(TODO: remove duplication of character token value completely and just use .raw instead)