Remove JS implementation and parser-feedback-tests
JS implementation was very useful for prototyping, but we don't keep it up-to-date anymore and Ragel seems to drop support for 3rd-party languages in 7.x branch.

If we ever want to get back and implement LazyHTML for other languages, we can always do so, but for now it's better to reduce unnecessary burden.